Feat: AdminCommandService - verifyTicketOwner
